test(appointment): DatePeriodTest 구현
